+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/setuptools/SetupToolsExtractor.java
@@ -1,73 +1,33 @@
 package com.synopsys.integration.detectable.detectables.setuptools;
 
 import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
 
 import org.apache.commons.lang3.StringUtils;
-import org.tomlj.TomlArray;
-import org.tomlj.TomlParseResult;
 
-import com.synopsys.integration.bdio.graph.BasicDependencyGraph;
 import com.synopsys.integration.bdio.graph.DependencyGraph;
-import com.synopsys.integration.bdio.model.Forge;
-import com.synopsys.integration.bdio.model.dependency.Dependency;
-import com.synopsys.integration.bdio.model.externalid.ExternalId;
-import com.synopsys.integration.bdio.model.externalid.ExternalIdFactory;
 import com.synopsys.integration.detectable.ExecutableTarget;
-import com.synopsys.integration.detectable.ExecutableUtils;
 import com.synopsys.integration.detectable.detectable.codelocation.CodeLocation;
-import com.synopsys.integration.detectable.detectable.executable.DetectableExecutableRunner;
-import com.synopsys.integration.detectable.detectable.util.TomlFileUtils;
 import com.synopsys.integration.detectable.detectables.setuptools.parse.SetupToolsParsedResult;
 import com.synopsys.integration.detectable.detectables.setuptools.parse.SetupToolsParser;
+import com.synopsys.integration.detectable.detectables.setuptools.transform.SetupToolsGraphTransformer;
 import com.synopsys.integration.detectable.extraction.Extraction;
 import com.synopsys.integration.detectable.extraction.Extraction.Builder;
-import com.synopsys.integration.executable.ExecutableRunnerException;
 
 public class SetupToolsExtractor {
     
-    private final DetectableExecutableRunner executableRunner;
-    
-    private ExternalIdFactory externalIdFactory;
+    private final SetupToolsGraphTransformer setupToolsTransformer;
 
-    public SetupToolsExtractor(DetectableExecutableRunner executableRunner, ExternalIdFactory externalIdFactory) {
-        this.executableRunner = executableRunner;
-        this.externalIdFactory = externalIdFactory;
+    public SetupToolsExtractor(SetupToolsGraphTransformer setupToolsTransformer) {
+        this.setupToolsTransformer = setupToolsTransformer;
     }
 
     public Extraction extract(File sourceDirectory, SetupToolsParser parser, ExecutableTarget pipExe) {
         try {
-            // TODO this belongs in a Parser class? Then can have a parser class for each 
-            // file type. Maybe I can have an interface or parent class so I can just pass in
-            // DependencyFile or something like that and call DependencyFile.parse
-            
             // TODO get direct dependencies from Toml file. Eventually we'll need to get these from cfg and py files
             // instead if those exist or have dependencies. 
             SetupToolsParsedResult parsedResult = parser.parse();
-
-            // TODO the graph stuff is usually in a "transformer" class, I probably don't need the
-            // model or result classes as this is pretty simple so far. Like NpmLockfileGraphTransformer.transform
-            DependencyGraph dependencyGraph = new BasicDependencyGraph();
             
-            if (pipExe != null) {
-                // Get dependencies by running pip show on each direct dependency
-                for (String directDependency : parsedResult.getDirectDependencies()) {
-                    parseShowDependency(pipExe, sourceDirectory, dependencyGraph, directDependency, null);
-                }
-            } else {
-                // Unable to determine transitive dependencies, add parsed dependencies directly
-                // to the root of the graph.
-                for (String directDependency : parsedResult.getDirectDependencies()) {
-                    Dependency currentDependency = entryToDependency(directDependency);
-                    dependencyGraph.addChildrenToRoot(currentDependency);
-                }
-            }
+            DependencyGraph dependencyGraph = setupToolsTransformer.transform(pipExe, parsedResult);
 
             CodeLocation codeLocation = new CodeLocation(dependencyGraph);
 
@@ -89,65 +49,5 @@ public class SetupToolsExtractor {
         } catch (Exception e) {
             return new Extraction.Builder().exception(e).build();
         }
-    }
-
-    private void parseShowDependency(ExecutableTarget pipExe, File sourceDirectory, DependencyGraph dependencyGraph, String dependencyToSearch, Dependency parentDependency) throws IOException, InterruptedException, ExecutableRunnerException {
-        List<String> rawShowOutput = runPipShow(sourceDirectory, pipExe, dependencyToSearch);
-        
-        Map<String, String> showOutput = parsePipShow(rawShowOutput);
-        
-        Dependency currentDependency = entryToDependency(dependencyToSearch, showOutput.get("Version"));
-        
-        if (parentDependency == null) {
-            // No parent, add to root of graph
-            dependencyGraph.addChildrenToRoot(currentDependency);
-        } else {
-            dependencyGraph.addChildWithParent(currentDependency, parentDependency);
-        }
-        
-        // See if we need to continue exploring this chain of dependencies
-        if (showOutput.containsKey("Requires")) {
-            String[] requiredPackages = showOutput.get("Requires").split(", ");
-            for (String requiredPackage : requiredPackages) {
-                parseShowDependency(pipExe, sourceDirectory, dependencyGraph, requiredPackage, currentDependency);
-            }
-        }
-    }
-
-    public List<String> runPipShow(File sourceDirectory, ExecutableTarget pipExe, String dependencyToSearch) throws ExecutableRunnerException {
-        // TODO somewhere need checks to blow up if no dependencies are found.
-        
-        List<String> pipArguments = new ArrayList<>();
-        pipArguments.add("show");
-        pipArguments.add(dependencyToSearch);
-        
-        return executableRunner.execute(ExecutableUtils.createFromTarget(sourceDirectory, pipExe, pipArguments)).getStandardOutputAsList();
-    }
-    
-    public Map<String, String> parsePipShow(List<String> rawShowOutput) {
-        Map<String, String> showOutput = new HashMap<>();
-        
-        for (String line : rawShowOutput) {
-            String[] parts = line.split(": ");
-            if (parts.length >= 2) {
-                String key = parts[0].trim();
-                String value = parts[1].trim();
-                if (key.equals("Version") || key.equals("Requires")) {
-                    showOutput.put(key, value);
-                }
-            }
-        }
-
-        return showOutput;
-    }
-    
-    private Dependency entryToDependency(String key) {
-        ExternalId externalId = externalIdFactory.createNameVersionExternalId(Forge.PYPI, key);
-        return new Dependency(externalId); 
-    }
-    
-    private Dependency entryToDependency(String key, String value) {
-        ExternalId externalId = externalIdFactory.createNameVersionExternalId(Forge.PYPI, key, value);
-        return new Dependency(externalId);
     }
 }
